+++ src/widgets/post/ui/consulting-response-sidebar/consulting-response-sidebar-additional-info-tab-view.tsx
