+++ src/widgets/post/ui/consulting-response-sidebar/consulting-response-sidebar-current-state-tab-view.tsx
