+++ src/widgets/post/ui/consulting-response-sidebar/consulting-response-sidebar-desired-style-tab-view.tsx
