+++ src/widgets/post/ui/consulting-response/consulting-response-price-and-comment-container.tsx
