+++ src/widgets/post/ui/consulting-response/consulting-response-recommend-style-container.tsx
